change price monthly

+++ workers-service/db/migration/000004_create_paid_monthly_table.up.sql
@@ -1,7 +1,7 @@
 CREATE TABLE IF NOT EXISTS paid_monthly (
     id SERIAL PRIMARY KEY,
     date DATE NOT NULL,
-    paid_price INT NOT NULL,
+    paid_price FLOAT NOT NULL,
     worker_id INT NOT NULL,
     created_at TIMESTAMP DEFAULT CURRENT_TIMESTAMP,
     updated_at TIMESTAMP DEFAULT CURRENT_TIMESTAMP,
